pyproject.toml
# ==============================================================================
# Файл конфигурации проекта для Poetry
# ==============================================================================

[tool.poetry]
name = "ml-foundry"
version = "0.1.0"
description = "Гибкий и воспроизводимый фреймворк для соревнований по машинному обучению"
authors = ["Artyom Konukhov <artemkonuhov008@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

# ==============================================================================
# --- Основные зависимости (для работы фреймворка) ---
# ==============================================================================
[tool.poetry.dependencies]
python = "^3.9" # 👈 Указывает, что проект совместим с Python 3.9 и выше, но не с 4.0

# --- Основа и Управление ---
hydra-core = "^1.3.2"           # Для управления конфигурациями
omegaconf = "^2.3.0"            # Используется Hydra для работы с конфигами
wandb = "^0.17.0"               # Для отслеживания экспериментов

# --- Обработка данных ---
pandas = "^2.2.2"
numpy = "^1.26.4"
scikit-learn = "^1.4.2"         # Для метрик, масштабирования, простых моделей
pyarrow = "^16.0.0"             # Для работы с форматом .parquet

# --- Модели ---
lightgbm = "^4.3.0"
xgboost = "^2.0.3"
catboost = "^1.2.5"

# --- Фиче-инжиниринг (Продвинутый) ---
gensim = "^4.3.2"               # Для классических эмбеддингов (Word2Vec, GloVe)
torch = "^2.3.0"                # Для нейронных сетей (Autoencoder, TabularNN)
transformers = "^4.40.1"        # Для трансформерных эмбеддингов
category-encoders = "^2.6.3"    # Для HashingEncoder и других
optuna = "^3.6.1"               # Для подбора гиперпараметров
optuna-integration = "^3.6.1"   # Для интеграции Optuna с W&B

# ==============================================================================
# --- Зависимости для разработки (не нужны для простого запуска) ---
# ==============================================================================
[tool.poetry.group.dev.dependencies]
# --- Для интерактивной работы и EDA ---
jupyterlab = "^4.1.6"
ipykernel = "^6.29.4"
matplotlib = "^3.8.4"
seaborn = "^0.13.2"

# --- Для качества кода и тестирования ---
pytest = "^8.2.0"               # Для юнит-тестирования
black = "^24.4.2"               # Для автоматического форматирования кода
flake8 = "^7.0.0"               # Для проверки стиля кода (линтер)
isort = "^5.13.2"               # Для сортировки импортов

# ==============================================================================
# --- Секция для сборки (обычно не изменяется) ---
# ==============================================================================
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.main.api"